+++ src/main/kotlin/com/itmo/se/recommendationservice/util/OffsetBasedPageRequest.kt
@@ -1,4 +1,4 @@
-package com.github.kuzznya.demo.util
+package com.itmo.se.recommendationservice.util
 
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
